+++ suppaftp/src/async_ftp.rs
@@ -21,24 +21,24 @@ pub mod async_std {
     pub type AsyncFtpStream = ImplAsyncFtpStream<AsyncNoTlsStream>;
 
     // -- export async secure (native-tls)
-    #[cfg(all(feature = "async-std", feature = "async-native-tls-std"))]
+    #[cfg(all(feature = "async-std", feature = "async-std-async-native-tls"))]
     #[cfg_attr(
         docsrs,
-        doc(cfg(all(feature = "async-std", feature = "async-native-tls-std")))
+        doc(cfg(all(feature = "async-std", feature = "async-std-async-native-tls")))
     )]
     pub use crate::async_ftp::async_std_ftp::AsyncNativeTlsConnector;
-    #[cfg(all(feature = "async-std", feature = "async-native-tls-std"))]
+    #[cfg(all(feature = "async-std", feature = "async-std-async-native-tls"))]
     #[cfg_attr(
         docsrs,
-        doc(cfg(all(feature = "async-std", feature = "async-native-tls-std")))
+        doc(cfg(all(feature = "async-std", feature = "async-std-async-native-tls")))
     )]
     use crate::async_ftp::async_std_ftp::AsyncNativeTlsStream;
     #[cfg(feature = "async-std")]
     #[cfg_attr(docsrs, doc(cfg(feature = "async-std")))]
     pub use crate::async_ftp::async_std_ftp::DataStream as AsyncDataStream;
 
-    #[cfg(feature = "async-native-tls-std")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "async-native-tls-std")))]
+    #[cfg(feature = "async-std-async-native-tls")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "async-std-async-native-tls")))]
     pub type AsyncNativeTlsFtpStream = ImplAsyncFtpStream<AsyncNativeTlsStream>;
     // -- export async secure (rustls)async-std-rustls
     #[cfg(all(feature = "async-std", feature = "async-std-rustls"))]
@@ -65,22 +65,22 @@ pub mod tokio {
     pub use super::tokio_ftp::{AsyncNoTlsStream, ImplAsyncFtpStream};
     pub type AsyncFtpStream = ImplAsyncFtpStream<AsyncNoTlsStream>;
 
-    #[cfg(all(feature = "tokio", feature = "async-native-tls-tokio"))]
+    #[cfg(all(feature = "tokio", feature = "tokio-async-native-tls"))]
     #[cfg_attr(
         docsrs,
-        doc(cfg(all(feature = "tokio", feature = "async-native-tls-tokio")))
+        doc(cfg(all(feature = "tokio", feature = "tokio-async-native-tls")))
     )]
     pub use super::tokio_ftp::AsyncNativeTlsConnector;
-    #[cfg(all(feature = "tokio", feature = "async-native-tls-tokio"))]
+    #[cfg(all(feature = "tokio", feature = "tokio-async-native-tls"))]
     #[cfg_attr(
         docsrs,
-        doc(cfg(all(feature = "tokio", feature = "async-native-tls-tokio")))
+        doc(cfg(all(feature = "tokio", feature = "tokio-async-native-tls")))
     )]
     use super::tokio_ftp::AsyncNativeTlsStream;
     pub use super::tokio_ftp::DataStream as AsyncDataStream;
 
-    #[cfg(feature = "async-native-tls-tokio")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "async-native-tls-tokio")))]
+    #[cfg(feature = "tokio-async-native-tls")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "tokio-async-native-tls")))]
     pub type AsyncNativeTlsFtpStream = ImplAsyncFtpStream<AsyncNativeTlsStream>;
 
     #[cfg(all(feature = "tokio", feature = "tokio-rustls"))]